requirements.txt
absl-py==2.3.1
agents==1.4.0
aiofiles==23.2.1
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.121.3
ffmpy==1.0.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.47.0
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
macholib @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.12.0
networkx==3.2.1
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pillow==10.4.0
pluggy==1.6.0
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
ruff==0.14.6
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.12.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==12.0
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0
